requirements.txt
Django==4.2.11
djangorestframework==3.14.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-filter==23.5
